consumer/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
psycopg2-binary==2.9.9
SQLAlchemy==2.0.32
kafka-python==2.0.2
email-validator>=2.1.0
